+++ plugins/schedule/bulk_upload/lib/model/kBulkUploadScheduleEventCsvJobData.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @package plugins.scheduleBulkUpload
+ * @subpackage model.data
+ */
+
+class kBulkUploadScheduleEventCsvJobData extends kBulkUploadScheduleEventJobData
+{
+	
+}